AP-8433: Rework Python preference page
So far, the Python preference page covered all parts of the MVC pattern in a monolithic fashion. Since, in the future, we want to offer the functionality of the preference page also via the Python node dialogs, those parts need to be separated to allow to easily interchange the preference page view (SWT) with the node dialog view (AWT).

This commit comprises the following changes:

- Separate model aspect of the preference page into config classes that are based on settings models (org.knime.python2.config package) as well as into PythonPreferences.java.
- Separate view aspect of the preference page into panel classes that each map to a config object (org.knime.python2.prefs package); PythonPreferencePage.java is a composition of these panels.
- Separate observation of model changes into PythonEnvironmentConfigObserver.java. Here, installation tests for changed environments are run. This somewhat corresponds to a controller; however, some controlling is still done in PythonPreferencePage.java since that's where Eclipse hands control over to us.
- Separate saving and loading of preferences (i.e., settings models) into persistor and storage classes (org.knime.pyton2.prefs package).

Note that this only prepares porting the preference page's functionality to the node dialogs. The actual implementation will be subject to future commits.